chapter05-4 예제2-2

+++ StudyCProgrammingFromBrother/src/chapter05.c
@@ -3,7 +3,7 @@
 int main(void) {
     int num1=3, num2=4;
     double divResult;
-    divResult=num1/num2;
+    divResult=(double)num1/num2;
     printf("나눗셈 결과: %f\n", divResult);
     return 0;
 }
